requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
dj-database-url==3.0.1
Django==5.1.6
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.28.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonpath-python==1.0.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==3.26.1
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
shippo==3.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
typing-inspect==0.9.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.2.1
whitenoise==6.6.0
